refactor(dubbo): deprecate service module, add framework file
